Refactor(performance): Split the parallel module
parallel.py held 587 lines, past the 400-line reviewability limit the
aislop code-quality gate applies.

Split it into a package at the same dotted path: models (the worker
and status enums plus the config/result records), aggregator (the
thread-safe ResultAggregator), worker_pool (the executor abstraction)
and processor (ParallelRepositoryProcessor and parallel_map). The
package __init__ re-exports the previous names.
