+++ src/performance/parallel/processor.py
@@ -1,303 +1,24 @@
 # SPDX-License-Identifier: Apache-2.0
-# SPDX-FileCopyrightText: 2025 The Linux Foundation
+# SPDX-FileCopyrightText: 2026 The Linux Foundation
 
 """
-Parallel repository processing module for the Repository Reporting System.
+Parallel repository processing coordination.
 
-This module provides utilities for processing multiple repositories in parallel
-using a worker pool architecture. It handles result aggregation, error management,
-and progress tracking across workers.
-
-Classes:
-    ParallelRepositoryProcessor: Main coordinator for parallel processing
-    WorkerPool: Worker lifecycle management
-    ResultAggregator: Thread-safe result collection
-    WorkerConfig: Configuration for parallel processing
-    ProcessingResult: Result from processing a repository
-
-Example:
-    >>> from src.performance import ParallelRepositoryProcessor
-    >>>
-    >>> processor = ParallelRepositoryProcessor(max_workers=4)
-    >>> repos = ['repo1', 'repo2', 'repo3']
-    >>> results = processor.process_repositories(repos, analyze_function)
-    >>> print(f"Processed {len(results.successful)} repositories")
+Drives the worker pool, dispatches per-repository work with optional
+retry and profiling, aggregates results, and exposes the simple
+``parallel_map`` convenience wrapper.
 """
 
-import multiprocessing
 import threading
 import time
 import traceback
 from collections.abc import Callable
-from concurrent.futures import Future, ProcessPoolExecutor, ThreadPoolExecutor, as_completed
-from dataclasses import dataclass, field
-from enum import Enum
+from concurrent.futures import Future, as_completed
 from typing import Any
 
-
-class WorkerType(Enum):
-    """Type of worker pool to use."""
-
-    THREAD = "thread"
-    PROCESS = "process"
-
-
-class ProcessingStatus(Enum):
-    """Status of a processing task."""
-
-    PENDING = "pending"
-    RUNNING = "running"
-    SUCCESS = "success"
-    FAILED = "failed"
-    TIMEOUT = "timeout"
-    CANCELLED = "cancelled"
-
-
-@dataclass
-class WorkerConfig:
-    """Configuration for parallel processing."""
-
-    max_workers: int = 4
-    worker_type: WorkerType = WorkerType.THREAD
-    worker_timeout: int = 300  # seconds
-    batch_size: int = 10
-    retry_on_failure: bool = False
-    max_retries: int = 2
-    stop_on_error: bool = False
-
-    def __post_init__(self):
-        """Validate configuration."""
-        if self.max_workers < 1:
-            raise ValueError(f"max_workers must be >= 1, got {self.max_workers}")
-        if self.max_workers > 64:
-            raise ValueError(f"max_workers must be <= 64, got {self.max_workers}")
-        if self.worker_timeout < 1:
-            raise ValueError(f"worker_timeout must be >= 1, got {self.worker_timeout}")
-        if self.batch_size < 1:
-            raise ValueError(f"batch_size must be >= 1, got {self.batch_size}")
-        if self.max_retries < 0:
-            raise ValueError(f"max_retries must be >= 0, got {self.max_retries}")
-
-    @staticmethod
-    def auto_detect_workers() -> int:
-        """
-        Auto-detect optimal number of workers.
-
-        Returns:
-            Recommended worker count
-        """
-        cpu_count = multiprocessing.cpu_count()
-        # Use CPU count for I/O-bound tasks (repository analysis is I/O heavy)
-        # Cap at 16 to avoid overwhelming system
-        return min(cpu_count, 16)
-
-
-@dataclass
-class ProcessingResult:
-    """Result from processing a single repository."""
-
-    item_id: str
-    status: ProcessingStatus
-    result: Any = None
-    error: str | None = None
-    error_traceback: str | None = None
-    start_time: float = 0.0
-    end_time: float = 0.0
-    worker_id: int | None = None
-    retry_count: int = 0
-
-    @property
-    def duration(self) -> float:
-        """Duration in seconds."""
-        return self.end_time - self.start_time if self.end_time > 0 else 0.0
-
-    @property
-    def is_success(self) -> bool:
-        """Check if processing was successful."""
-        return self.status == ProcessingStatus.SUCCESS
-
-    @property
-    def is_failure(self) -> bool:
-        """Check if processing failed."""
-        return self.status in (ProcessingStatus.FAILED, ProcessingStatus.TIMEOUT)
-
-
-@dataclass
-class AggregatedResults:
-    """Aggregated results from parallel processing."""
-
-    total: int
-    successful: list[ProcessingResult] = field(default_factory=list)
-    failed: list[ProcessingResult] = field(default_factory=list)
-    results: dict[str, Any] = field(default_factory=dict)
-    errors: dict[str, str] = field(default_factory=dict)
-    total_duration: float = 0.0
-
-    @property
-    def success_count(self) -> int:
-        """Number of successful results."""
-        return len(self.successful)
-
-    @property
-    def failure_count(self) -> int:
-        """Number of failed results."""
-        return len(self.failed)
-
-    @property
-    def success_rate(self) -> float:
-        """Success rate as percentage."""
-        return (self.success_count / self.total * 100) if self.total > 0 else 0.0
-
-    @property
-    def avg_duration(self) -> float:
-        """Average processing duration."""
-        if not self.successful:
-            return 0.0
-        return sum(r.duration for r in self.successful) / len(self.successful)
-
-
-class ResultAggregator:
-    """
-    Thread-safe result aggregator for collecting results from workers.
-
-    Example:
-        >>> aggregator = ResultAggregator(total_items=10)
-        >>> aggregator.add_result(ProcessingResult(...))
-        >>> results = aggregator.get_results()
-    """
-
-    def __init__(self, total_items: int):
-        """
-        Initialize result aggregator.
-
-        Args:
-            total_items: Total number of items to process
-        """
-        self.total_items = total_items
-        self.lock = threading.Lock()
-        self.results: list[ProcessingResult] = []
-        self.completed_count = 0
-        self.start_time = time.perf_counter()
-
-    def add_result(self, result: ProcessingResult):
-        """
-        Add a result (thread-safe).
-
-        Args:
-            result: Processing result to add
-        """
-        with self.lock:
-            self.results.append(result)
-            self.completed_count += 1
-
-    def get_progress(self) -> tuple[int, int]:
-        """
-        Get current progress.
-
-        Returns:
-            Tuple of (completed, total)
-        """
-        with self.lock:
-            return (self.completed_count, self.total_items)
-
-    def get_results(self) -> AggregatedResults:
-        """
-        Get aggregated results.
-
-        Returns:
-            AggregatedResults with all data
-        """
-        with self.lock:
-            end_time = time.perf_counter()
-
-            successful = [r for r in self.results if r.is_success]
-            failed = [r for r in self.results if r.is_failure]
-
-            results_dict = {r.item_id: r.result for r in successful if r.result is not None}
-            errors_dict = {r.item_id: r.error for r in failed if r.error is not None}
-
-            return AggregatedResults(
-                total=self.total_items,
-                successful=successful,
-                failed=failed,
-                results=results_dict,
-                errors=errors_dict,
-                total_duration=end_time - self.start_time,
-            )
-
-
-class WorkerPool:
-    """
-    Worker pool for managing parallel execution.
-
-    Example:
-        >>> with WorkerPool(max_workers=4, worker_type=WorkerType.THREAD) as pool:
-        ...     results = pool.map(func, items)
-    """
-
-    def __init__(
-        self,
-        max_workers: int = 4,
-        worker_type: WorkerType = WorkerType.THREAD,
-        worker_timeout: int = 300,
-    ):
-        """
-        Initialize worker pool.
-
-        Args:
-            max_workers: Maximum number of workers
-            worker_type: Type of workers (thread or process)
-            worker_timeout: Timeout per task in seconds
-        """
-        self.max_workers = max_workers
-        self.worker_type = worker_type
-        self.worker_timeout = worker_timeout
-        self.executor: ThreadPoolExecutor | ProcessPoolExecutor | None = None
-
-    def __enter__(self):
-        """Enter context manager."""
-        if self.worker_type == WorkerType.THREAD:
-            self.executor = ThreadPoolExecutor(max_workers=self.max_workers)
-        else:
-            self.executor = ProcessPoolExecutor(max_workers=self.max_workers)
-        return self
-
-    def __exit__(self, exc_type, exc_val, exc_tb):
-        """Exit context manager."""
-        if self.executor:
-            self.executor.shutdown(wait=True)
-        return False
-
-    def submit(self, func: Callable[..., Any], *args: Any, **kwargs: Any) -> Future[Any]:
-        """
-        Submit a task to the pool.
-
-        Args:
-            func: Function to execute
-            *args: Positional arguments
-            **kwargs: Keyword arguments
-
-        Returns:
-            Future object
-        """
-        if not self.executor:
-            raise RuntimeError("WorkerPool must be used as context manager")
-        return self.executor.submit(func, *args, **kwargs)
-
-    def map(self, func: Callable[..., Any], items: list[Any]) -> list[Any]:
-        """
-        Map function over items in parallel.
-
-        Args:
-            func: Function to apply
-            items: Items to process
-
-        Returns:
-            List of results
-        """
-        if not self.executor:
-            raise RuntimeError("WorkerPool must be used as context manager")
-        return list(self.executor.map(func, items, timeout=self.worker_timeout))
+from .aggregator import ResultAggregator
+from .models import AggregatedResults, ProcessingResult, ProcessingStatus, WorkerConfig
+from .worker_pool import WorkerPool
 
 
 class ParallelRepositoryProcessor:
